+++ hotrod-project/hotrod-abstract-plugin/src/main/java/org/hotrod/plugin/ExportColumnsToXLSXOperation.java
@@ -28,12 +28,12 @@ public class ExportColumnsToXLSXOperation extends AbstractExportColumnsOperation
   private int line;
   private int col;
 
-  public ExportColumnsToXLSXOperation(final File baseDir, final String configfilename, final String generator,
-      final String localproperties, final String jdbcdriverclass, final String jdbcurl, final String jdbcusername,
-      final String jdbcpassword, final String jdbccatalog, final String jdbcschema, final String facets,
-      final String display, final String exportfilename) {
-    super(baseDir, configfilename, generator, localproperties, jdbcdriverclass, jdbcurl, jdbcusername, jdbcpassword,
-        jdbccatalog, jdbcschema, facets, display, exportfilename);
+  public ExportColumnsToXLSXOperation(final File baseDir, final String configfilename, final String localproperties,
+      final String jdbcdriverclass, final String jdbcurl, final String jdbcusername, final String jdbcpassword,
+      final String jdbccatalog, final String jdbcschema, final String facets, final String display,
+      final String exportfilename) {
+    super(baseDir, configfilename, localproperties, jdbcdriverclass, jdbcurl, jdbcusername, jdbcpassword, jdbccatalog,
+        jdbcschema, facets, display, exportfilename);
     log.debug("init");
   }