+++ hotrod-project/hotrod-ant-plugin/src/main/java/org/hotrod/plugin/ant/ExportColumnsToXLSXAntTask.java
@@ -13,7 +13,6 @@ public class ExportColumnsToXLSXAntTask extends Task {
   private static transient final Logger log = LogManager.getLogger(ExportColumnsToXLSXAntTask.class);
 
   private String configfile = null;
-  private String generator = null;
   private String localproperties = null;
 
   private String jdbcdriverclass = null;
@@ -31,7 +30,7 @@ public class ExportColumnsToXLSXAntTask extends Task {
   public void execute() {
     log.debug("init");
 
-    ExportColumnsToXLSXOperation op = new ExportColumnsToXLSXOperation(new File("."), this.configfile, this.generator,
+    ExportColumnsToXLSXOperation op = new ExportColumnsToXLSXOperation(new File("."), this.configfile,
         this.localproperties, this.jdbcdriverclass, this.jdbcurl, this.jdbcusername, this.jdbcpassword,
         this.jdbccatalog, this.jdbcschema, this.display, this.facets, this.exportfilename);
 
@@ -47,10 +46,6 @@ public class ExportColumnsToXLSXAntTask extends Task {
 
   public void setConfigfile(final String configfile) {
     this.configfile = configfile;
-  }
-
-  public void setGenerator(final String generator) {
-    this.generator = generator;
   }
 
   public void setLocalproperties(String localproperties) {