+++ hotrod-project/hotrod-generator/src/main/java/org/hotrod/config/SelectGenerationTag.java
@@ -22,7 +22,7 @@ public class SelectGenerationTag extends AbstractConfigurationTag {
 
   private static final Logger log = LogManager.getLogger(SelectGenerationTag.class);
 
-  public static final SelectStrategy DEFAULT_STRATEGY = SelectStrategy.CREATE_VIEW;
+  public static final SelectStrategy DEFAULT_STRATEGY = SelectStrategy.RESULT_SET;
   public static final String DEFAULT_TEMP_VIEW_NAME = "hotrodtempview";
 
   private static final String VIEW_NAME_PATTERN = "[a-zA-Z][a-zA-Z0-9]*";