+++ hotrod-project/docs/version-history.md
@@ -6,6 +6,8 @@ Libraries for these versions can be downloaded from any Maven Repository mirror.
 - Support for multiple datasources.
 - LiveSQL now returns `List<Row>` and `Cursor<Row>` instead of `List<Map<String, Object>>` and `Cursor<Map<String, Object>>`.
 - `propertiesChangeLog` is now removed from VOs to facilitate quick Spring prototyping.
+- The `<select-generation>` tag now defaults to `result-set`, so it can be omitted in the configuration file.
+- The `<classic-fk-navigation>` tag is not enabled by default, so it can be omitted in the configuration file.
 
 ## 3.4.8
 - Adding support for H2 version 2.x.